+++ lib/openapi_first.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'yaml'
+require 'multi_json'
 require 'json_refs'
 require_relative 'openapi_first/errors'
 require_relative 'openapi_first/configuration'
@@ -8,8 +9,8 @@ require_relative 'openapi_first/plugins'
 require_relative 'openapi_first/definition'
 require_relative 'openapi_first/version'
 require_relative 'openapi_first/error_response'
-require_relative 'openapi_first/request_validation'
-require_relative 'openapi_first/response_validation'
+require_relative 'openapi_first/middlewares/response_validation'
+require_relative 'openapi_first/middlewares/request_validation'
 
 module OpenapiFirst
   extend Plugins
@@ -24,15 +25,27 @@ module OpenapiFirst
     end
   end
 
-  # An instance of RuntimeRequest
+  # Key in rack to find instance of RuntimeRequest
   REQUEST = 'openapi.request'
 
   def self.load(spec_path, only: nil)
-    resolved = Dir.chdir(File.dirname(spec_path)) do
-      content = YAML.load_file(File.basename(spec_path))
-      JsonRefs.call(content, resolve_local_ref: true, resolve_file_ref: true)
-    end
+    resolved = Bundle.resolve(spec_path)
     resolved['paths'].filter!(&->(key, _) { only.call(key) }) if only
     Definition.new(resolved, spec_path)
+  end
+
+  module Bundle
+    def self.resolve(spec_path)
+      Dir.chdir(File.dirname(spec_path)) do
+        content = load_file(File.basename(spec_path))
+        JsonRefs.call(content, resolve_local_ref: true, resolve_file_ref: true)
+      end
+    end
+
+    def self.load_file(spec_path)
+      return MultiJson.load(File.read(spec_path)) if File.extname(spec_path) == '.json'
+
+      YAML.unsafe_load_file(spec_path)
+    end
   end
 end