+++ lib/openapi_first/runtime_request.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'forwardable'
-require 'openapi_parameters'
 require_relative 'runtime_response'
 require_relative 'body_parser'
 require_relative 'request_validation/validator'
@@ -18,7 +17,7 @@ module OpenapiFirst
       @original_path_params = path_params
     end
 
-    def_delegators :@request, :content_type
+    def_delegators :@request, :content_type, :media_type
     def_delegators :@operation, :operation_id
 
     def known?
@@ -35,11 +34,11 @@ module OpenapiFirst
 
     # Merged path and query parameters
     def params
-      @params ||= query.merge(path_params)
+      @params ||= query.merge(path_parameters)
     end
 
-    def path_params
-      @path_params ||=
+    def path_parameters
+      @path_parameters ||=
         operation.path_parameters&.unpack(@original_path_params) || {}
     end
 
@@ -47,6 +46,8 @@ module OpenapiFirst
       @query ||=
         operation.query_parameters&.unpack(request.env) || {}
     end
+
+    alias query_parameters query
 
     def headers
       @headers ||=
@@ -61,6 +62,7 @@ module OpenapiFirst
     def body
       @body ||= BodyParser.new.parse(request, request.media_type)
     end
+    alias parsed_body body
 
     def validate
       RequestValidation::Validator.new(operation).validate(self)