+++ spec/plugins/jsonapi/error_response_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe OpenapiFirst::Plugins::Jsonapi::ErrorResponse do
     context 'when validation_result is nil' do
       specify do
         error = described_class.new(
-          failure: OpenapiFirst::RequestValidation::Failure.new(
+          failure: OpenapiFirst::Failure.new(
             :invalid_body
           )
         )
@@ -41,7 +41,7 @@ RSpec.describe OpenapiFirst::Plugins::Jsonapi::ErrorResponse do
         data = { 'data' => { 'name' => 21, 'numberOfLegs' => 'four' } }
         validation_result = OpenapiFirst::Schema.new(schema, openapi_version: '3.1').validate(data)
         error = described_class.new(
-          failure: OpenapiFirst::RequestValidation::Failure.new(
+          failure: OpenapiFirst::Failure.new(
             :invalid_body,
             errors: validation_result.errors
           )