+++ spec/request_validation/cookie_parameter_validation_spec.rb
@@ -10,7 +10,8 @@ RSpec.describe 'Cookie Parameter validation' do
 
   let(:app) do
     Rack::Builder.app do
-      use OpenapiFirst::RequestValidation, spec: File.expand_path('../data/cookie-parameter-validation.yaml', __dir__)
+      use OpenapiFirst::Middlewares::RequestValidation,
+          spec: File.expand_path('../data/cookie-parameter-validation.yaml', __dir__)
       run lambda { |_env|
         Rack::Response.new('hello', 200).finish
       }
@@ -50,8 +51,8 @@ RSpec.describe 'Cookie Parameter validation' do
       let(:app) do
         Rack::Builder.app do
           spec_file = File.expand_path('../data/cookie-parameter-validation.yaml', __dir__)
-          use OpenapiFirst::RequestValidation, raise_error: true,
-                                               spec: spec_file
+          use OpenapiFirst::Middlewares::RequestValidation, raise_error: true,
+                                                            spec: spec_file
           run lambda { |_env|
             Rack::Response.new('hello', 200).finish
           }
@@ -61,7 +62,7 @@ RSpec.describe 'Cookie Parameter validation' do
       it 'returns 400 if cookie parameter is invalid' do
         expect do
           get '/'
-        end.to raise_error OpenapiFirst::RequestInvalidError, /^Cookie value invalid: \w+/
+        end.to raise_error OpenapiFirst::RequestInvalidError, /^Cookie value is invalid: \w+/
       end
     end
   end