+++ spec/request_validation/header_parameter_validation_spec.rb
@@ -10,7 +10,8 @@ RSpec.describe 'Header Parameter validation' do
 
   let(:app) do
     Rack::Builder.app do
-      use OpenapiFirst::RequestValidation, spec: File.expand_path('../data/header-parameter-validation.yaml', __dir__)
+      use OpenapiFirst::Middlewares::RequestValidation,
+          spec: File.expand_path('../data/header-parameter-validation.yaml', __dir__)
       run lambda { |_env|
         Rack::Response.new('hello', 200).finish
       }
@@ -47,8 +48,8 @@ RSpec.describe 'Header Parameter validation' do
       let(:app) do
         Rack::Builder.app do
           spec_file = File.expand_path('../data/header-parameter-validation.yaml', __dir__)
-          use OpenapiFirst::RequestValidation, raise_error: true,
-                                               spec: spec_file
+          use OpenapiFirst::Middlewares::RequestValidation, raise_error: true,
+                                                            spec: spec_file
           run lambda { |_env|
             Rack::Response.new('hello', 200).finish
           }
@@ -60,7 +61,7 @@ RSpec.describe 'Header Parameter validation' do
         expect do
           get '/pets'
         end.to raise_error OpenapiFirst::RequestInvalidError,
-                           'Header parameter invalid: value at `/Accept-Version` is not an integer'
+                           'Request header is invalid: value at `/Accept-Version` is not an integer'
       end
     end
   end